Improve root Makefile to make use of individual Makefiles
`make package` on the root directory will invoke the individual make
on each directory, which will allow for customized build behaviors
per-app.

If an individual app is already built, it won't be rebuilt. We are
providing a `cleanall` target to remove them too. `make clean` will
only remove the package dir from the top but not the individual ones.

Also, added missing individual Makefile for annex app.